Add etc/sofa.ini configuration file
This file is generated by CMake. It is written in a basic ini style, and
contains only the paths to the share/ and examples/ directories.

For now, it always points to the source tree, but the idea is that once
we can install Sofa, it will point to the installed directories.


Former-commit-id: 94d44e803c5a8cb019fef830058e65fafa1a8b43